+++ internal/terraform/node_resource_plan_orphan.go
@@ -24,6 +24,14 @@ type NodePlannableResourceInstanceOrphan struct {
 	// skipPlanChanges indicates we should skip trying to plan change actions
 	// for any instances.
 	skipPlanChanges bool
+
+	// forgetResources lists resources that should not be destroyed, only removed
+	// from state.
+	forgetResources []addrs.ConfigResource
+
+	// forgetModules lists modules that should not be destroyed, only removed
+	// from state.
+	forgetModules []addrs.Module
 }
 
 var (
@@ -132,14 +140,31 @@ func (n *NodePlannableResourceInstanceOrphan) managedResourceExecute(ctx EvalCon
 		return diags.Append(n.writeResourceInstanceState(ctx, oldState, workingState))
 	}
 
-	var change *plans.ResourceInstanceChange
-	change, destroyPlanDiags := n.planDestroy(ctx, oldState, "")
-	diags = diags.Append(destroyPlanDiags)
-	if diags.HasErrors() {
-		return diags
+	var forget bool
+	for _, ft := range n.forgetResources {
+		if ft.Equal(n.ResourceAddr()) {
+			forget = true
+		}
 	}
+	for _, fm := range n.forgetModules {
+		if fm.Equal(n.Addr.Module.Module()) {
+			forget = true
+		}
+	}
+	var change *plans.ResourceInstanceChange
+	var pDiags tfdiags.Diagnostics
+	if forget {
+		change, pDiags = n.planForget(ctx, oldState, "")
+		diags = diags.Append(pDiags)
+	} else {
+		change, pDiags = n.planDestroy(ctx, oldState, "")
+		diags = diags.Append(pDiags)
+		if diags.HasErrors() {
+			return diags
+		}
 
-	diags = diags.Append(n.checkPreventDestroy(change))
+		diags = diags.Append(n.checkPreventDestroy(change))
+	}
 	if diags.HasErrors() {
 		return diags
 	}